Tarea 3.8 comenzada

+++ src/main/java/org/iesvdm/dao/ComercialDAOImpl.java
@@ -46,7 +46,7 @@ public class ComercialDAOImpl implements ComercialDAO {
 			ps.setString(idx++, comercial.getNombre());
 			ps.setString(idx++, comercial.getApellido1());
 			ps.setString(idx++, comercial.getApellido2());
-			ps.setFloat(idx++, comercial.getComision());
+			ps.setBigDecimal(idx++, comercial.getComision());
 			return ps;
 		}, keyHolder);
 
@@ -70,7 +70,7 @@ public class ComercialDAOImpl implements ComercialDAO {
 
 		List<Comercial> listComercial = jdbcTemplate.query("SELECT * FROM comercial",
 				(rs, rowNum) -> new Comercial(rs.getInt("id"), rs.getString("nombre"), rs.getString("apellido1"),
-						rs.getString("apellido2"), rs.getFloat("comisión"))
+						rs.getString("apellido2"), rs.getBigDecimal("comisión"))
 
 		);
 
@@ -88,7 +88,7 @@ public class ComercialDAOImpl implements ComercialDAO {
             						 						rs.getString("nombre"),
             						 						rs.getString("apellido1"),
             						 						rs.getString("apellido2"),
-            						 						rs.getFloat("comisión")) 
+            						 						rs.getBigDecimal("comisión")) 
 								, id
 								);
